Fixing Siddhi aggregation functions return incorrect values when multiple invocations are made
https://github.com/wso2/siddhi/issues/750
