+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/FindStoreQueryRuntime.java
@@ -47,29 +47,33 @@ public class FindStoreQueryRuntime implements StoreQueryRuntime {
     private AggregationRuntime aggregation;
     private QuerySelector selector;
     private StateEventPool stateEventPool;
+    private MetaStreamEvent metaStreamEvent;
 
     public FindStoreQueryRuntime(Table table, CompiledCondition compiledCondition, String queryName,
-                                 MetaStreamEvent.EventType eventType) {
+                                 MetaStreamEvent metaStreamEvent) {
         this.table = table;
         this.compiledCondition = compiledCondition;
         this.queryName = queryName;
-        this.eventType = eventType;
+        this.eventType = metaStreamEvent.getEventType();
+        this.metaStreamEvent = metaStreamEvent;
     }
 
     public FindStoreQueryRuntime(Window window, CompiledCondition compiledCondition, String queryName,
-                                 MetaStreamEvent.EventType eventType) {
+                                 MetaStreamEvent metaStreamEvent) {
         this.window = window;
         this.compiledCondition = compiledCondition;
         this.queryName = queryName;
-        this.eventType = eventType;
+        this.eventType = metaStreamEvent.getEventType();
+        this.metaStreamEvent = metaStreamEvent;
     }
 
     public FindStoreQueryRuntime(AggregationRuntime aggregation, CompiledCondition compiledCondition, String queryName,
-                                 MetaStreamEvent.EventType eventType) {
+                                 MetaStreamEvent metaStreamEvent) {
         this.aggregation = aggregation;
         this.compiledCondition = compiledCondition;
         this.queryName = queryName;
-        this.eventType = eventType;
+        this.eventType = metaStreamEvent.getEventType();
+        this.metaStreamEvent = metaStreamEvent;
     }
 
     @Override
@@ -110,6 +114,31 @@ public class FindStoreQueryRuntime implements StoreQueryRuntime {
         } catch (Throwable t) {
             throw new StoreQueryRuntimeException("Error executing '" + queryName + "', " + t.getMessage(), t);
         }
+    }
+
+    @Override
+    public void reset() {
+        if (selector != null) {
+            selector.process(generateResetComplexEventChunk(metaStreamEvent));
+        }
+    }
+
+    private ComplexEventChunk<ComplexEvent> generateResetComplexEventChunk(MetaStreamEvent metaStreamEvent) {
+        StreamEvent streamEvent = new StreamEvent(metaStreamEvent.getBeforeWindowData().size(),
+                metaStreamEvent.getOnAfterWindowData().size(), metaStreamEvent.getOutputData().size());
+        streamEvent.setType(ComplexEvent.Type.RESET);
+
+        StateEvent stateEvent = stateEventPool.borrowEvent();
+        if (eventType == MetaStreamEvent.EventType.AGGREGATE) {
+            stateEvent.addEvent(1, streamEvent);
+        } else {
+            stateEvent.addEvent(0, streamEvent);
+        }
+        stateEvent.setType(ComplexEvent.Type.RESET);
+
+        ComplexEventChunk<ComplexEvent> complexEventChunk = new ComplexEventChunk<>(true);
+        complexEventChunk.add(stateEvent);
+        return complexEventChunk;
     }
 
     public void setStateEventPool(StateEventPool stateEventPool) {
