+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/query/SelectStoreQueryRuntime.java
@@ -65,4 +65,9 @@ public class SelectStoreQueryRuntime implements StoreQueryRuntime {
         }
     }
 
+    @Override
+    public void reset() {
+
+    }
+
 }
