+++ modules/siddhi-core/src/test/java/org/wso2/siddhi/core/store/StoreQueryTableTestCase.java
@@ -360,4 +360,90 @@ public class StoreQueryTableTestCase {
         AssertJUnit.assertEquals(55.6F, events[0].getData()[1]);
         AssertJUnit.assertEquals(75.6f, events[1].getData()[1]);
     }
+
+
+    @Test
+    public void test10() throws InterruptedException {
+        log.info("Test10 table");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String streams = "" +
+                "define stream StockStream (symbol string, price float, volume long);" +
+                "@PrimaryKey('symbol') " +
+                "define table StockTable (symbol string, price float, volume long); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from StockStream " +
+                "insert into StockTable ;";
+
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
+
+        siddhiAppRuntime.start();
+
+        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
+        stockStream.send(new Object[]{"IBM", 75.6f, 100L});
+        stockStream.send(new Object[]{"WSO2", 57.6f, 100L});
+        Thread.sleep(500);
+
+        String storeQuery = "" +
+                "from StockTable " +
+                "on volume > 10 " +
+                "select symbol, price, sum(volume) as totalVolume ";
+        Event[] events = siddhiAppRuntime.query(storeQuery);
+        EventPrinter.print(events);
+        AssertJUnit.assertEquals(1, events.length);
+        AssertJUnit.assertEquals(200L, events[0].getData()[2]);
+
+        events = siddhiAppRuntime.query(storeQuery);
+        EventPrinter.print(events);
+        AssertJUnit.assertEquals(1, events.length);
+        AssertJUnit.assertEquals(200L, events[0].getData()[2]);
+    }
+
+    @Test
+    public void test11() throws InterruptedException {
+        log.info("Test10 table");
+
+        SiddhiManager siddhiManager = new SiddhiManager();
+
+        String streams = "" +
+                "define stream StockStream (symbol string, price float, volume long);" +
+                "@PrimaryKey('symbol') " +
+                "define table StockTable (symbol string, price float, volume long); ";
+        String query = "" +
+                "@info(name = 'query1') " +
+                "from StockStream " +
+                "insert into StockTable ;";
+
+        SiddhiAppRuntime siddhiAppRuntime = siddhiManager.createSiddhiAppRuntime(streams + query);
+
+        InputHandler stockStream = siddhiAppRuntime.getInputHandler("StockStream");
+
+        siddhiAppRuntime.start();
+
+        stockStream.send(new Object[]{"WSO2", 55.6f, 100L});
+        stockStream.send(new Object[]{"IBM", 75.6f, 100L});
+        stockStream.send(new Object[]{"WSO2", 57.6f, 100L});
+        Thread.sleep(500);
+
+        String storeQuery = "" +
+                "from StockTable " +
+                "on volume > 10 " +
+                "select symbol, price, sum(volume) as totalVolume " +
+                "group by symbol ";
+        Event[] events = siddhiAppRuntime.query(storeQuery);
+        EventPrinter.print(events);
+        AssertJUnit.assertEquals(2, events.length);
+        AssertJUnit.assertEquals(100L, events[0].getData()[2]);
+        AssertJUnit.assertEquals(100L, events[1].getData()[2]);
+
+        events = siddhiAppRuntime.query(storeQuery);
+        EventPrinter.print(events);
+        AssertJUnit.assertEquals(2, events.length);
+        AssertJUnit.assertEquals(100L, events[0].getData()[2]);
+        AssertJUnit.assertEquals(100L, events[1].getData()[2]);
+    }
 }
